requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.23
pillow>=10.0
tqdm>=4.66
scikit-learn>=1.3
matplotlib>=3.7
opencv-python>=4.8
streamlit>=1.28
pyyaml>=6.0
